+++ src/App.tsx
@@ -12,6 +12,7 @@ import CourseUpdates, { CourseUpdateType, getCoursesDifference, returnColor } fr
 import Degree from './domain/Degree'
 
 import saveToExcel from './utils/excel'
+import getCalendar from './utils/calendar-generator'
 
 import i18next from 'i18next'
 import withStyles, { CreateCSSProperties } from '@material-ui/core/styles/withStyles'
@@ -35,7 +36,6 @@ import ToggleButtonGroup from '@material-ui/lab/ToggleButtonGroup'
 import Backdrop from '@material-ui/core/Backdrop'
 import CircularProgress from '@material-ui/core/CircularProgress'
 import Divider from '@material-ui/core/Divider'
-import { exportComponentAsPNG } from 'react-component-export-image'
 import downloadAsImage from './utils/save-as-image'
 import Snackbar from '@material-ui/core/Snackbar'
 import Link from '@material-ui/core/Link'
@@ -75,6 +75,7 @@ class App extends React.Component <{
 		hasAlert: false as boolean,
 		classesDialog: false,
 		warningDialog: false,
+		changelogDialog: false,
 		loading: true as boolean,
 		lang: i18next.options.lng as string,
 		darkMode: false
@@ -135,11 +136,11 @@ class App extends React.Component <{
 			loading: false
 		})
 
-		// Set warning of shifts with degrees that aren't the ones selected
-		const isWarned = this.cookies.get('warning-shift-degrees')
+		// Set warning with all notices
+		const isWarned = this.cookies.get('warning')
 		if (isWarned !== 'true') {
 			this.setWarningShiftDegrees()
-			this.cookies.set('warning-shift-degrees', 'true', { maxAge: 60*60*24 })
+			this.cookies.set('warning', 'true', { maxAge: 60*60*24 })
 		}
 	}
 
@@ -519,6 +520,12 @@ class App extends React.Component <{
 		this.showAlert(i18next.t('alert.schedule-to-excel'), 'success')
 	}
 
+	downloadCalendar(): void {
+		getCalendar(this.state.selectedShifts)
+
+		this.showAlert(i18next.t('alert.calendar-obtained'), 'success')
+	}
+
 	render(): ReactNode {
 		const classes = this.props.classes
 
@@ -671,6 +678,16 @@ class App extends React.Component <{
 													<Icon>download</Icon>
 												</IconButton>
 											</Tooltip>
+											<Tooltip title={i18next.t('schedule-selected.actions.get-calendar') as string}>
+												<IconButton
+													disabled={this.state.selectedShifts.length === 0}
+													color='inherit'
+													onClick={() => {this.downloadCalendar()}}
+													component="span"
+												>
+													<Icon>event</Icon>
+												</IconButton>
+											</Tooltip>
 											<Tooltip title={i18next.t('schedule-selected.actions.clear-schedule') as string}>
 												<IconButton
 													disabled={this.state.selectedShifts.length === 0}
@@ -697,6 +714,14 @@ class App extends React.Component <{
 										>{i18next.t('footer.support-button.content') as string}
 										</Button>
 									</Link>
+								</Tooltip>
+								<Tooltip title={i18next.t('footer.changelog-button.tooltip') as string} style={{marginLeft: '8px'}}>
+									<Button color='default' variant='outlined'
+										startIcon={<Icon>new_releases</Icon>}
+										size='small'
+										onClick={() => {this.setState({changelogDialog: true})}}
+									>{i18next.t('footer.changelog-button.content') as string}
+									</Button>
 								</Tooltip>
 								<div className={classes.grow as string} />
 								<Tooltip title={i18next.t('footer.repository.tooltip') as string}>
@@ -751,7 +776,15 @@ class App extends React.Component <{
 								<Button onClick={() => {this.warningContinue(); this.setState({warningDialog: false})}} color="primary">{i18next.t('warning.actions.continue') as string}</Button>
 								{/* <Button onClick={() => {this.setState({warningDialog: false})}} color="primary">{i18next.t('warning.actions.back') as string}</Button> */}
 							</DialogActions>
-						</Dialog>				
+						</Dialog>	
+						<Dialog open={this.state.changelogDialog}>
+							<DialogTitle>{i18next.t('changelog-dialog.title') as string}</DialogTitle>
+							<DialogContent style={{whiteSpace: 'pre-line'}}>{(i18next.t('changelog-dialog.content', {returnObjects: true}) as string[]).join('\n\n')}</DialogContent>
+							<DialogActions>
+								<div />
+								<Button onClick={() => {this.setState({changelogDialog: false})}} color="primary">{i18next.t('changelog-dialog.actions.back') as string}</Button>
+							</DialogActions>
+						</Dialog>			
 					</div>
 				</div>
 			</ThemeProvider>