+++ src/domain/Lesson.tsx
@@ -8,18 +8,22 @@ export default class Lesson implements Comparable {
 	type: string
 	startTime: string
 	endTime: string
+	date: string
 	daysOfWeek: number[]
 	color: string
 	id: string
-	minutes: number // in minutes
+	minutes: number
 	occupation: ShiftOccupation
 	courseUrl: string
+	courseName: string
+	room: string
 
 	// eslint-disable-next-line
 	constructor(obj: Record<string, any>) {
 		this.daysOfWeek = [obj.dayOfWeek]
 		this.startTime = obj.start
 		this.endTime = obj.end
+		this.date = obj.date
 		const dateStart = new Date('2021/02/17 ' + obj.start), dateEnd = new Date('2021/02/17 ' + obj.end)
 		this.minutes = (dateEnd.getTime() - dateStart.getTime()) / ( 1000 * 60 )
 		this.color = obj.color
@@ -27,6 +31,8 @@ export default class Lesson implements Comparable {
 		this.id = obj.shiftName
 		this.occupation = obj.occupation
 		this.courseUrl = obj.url
+		this.courseName = obj.courseName
+		this.room = obj.room
 		this.title = `${obj.acronym} - ${obj.shiftId}\n${ (obj.campus !== undefined) ? obj.campus : '' }`
 		this.exportedTitle = `${obj.acronym} - ${obj.shiftId} @ ${obj.room !== undefined ? obj.room : '' }`
 	}