+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.0] - 02/03/2021
+
+### Added
+
+- Allow download of calendar with the selected shifts
+- Add dialog with changelog
+
+### Fixed
+
+- React-scripts (immer) vulnerability
+- Excel file without filename
+
 ## [1.4.0] - 01/03/2021
 
 ### Added